+++ CHANGES.rst
@@ -6,6 +6,8 @@ Changelog of threedi-modelchecker
 2.14.2 (unreleased)
 -------------------
 
+- Check if tables related to use_* settings in model_settings and simulation_template settings are populated
+- Warn if tables related to use_* settings in model_settings and simulation_template settings are populated while use_* settings is false
 - Add test for check descriptions.
 - Bump minimal python version to 3.9
 - Collect all foreign key checks and give them a uniform error or warning (0001)